Refactor: Update dotenv loading and enhance error handling in authentication
